separate structs into separate files
